refactor(svelte): split interaction component integration test monolith
Move shared fixtures to interaction-harness and split the 1619-line
interaction.test.ts by top-level describe into focused suites under
tests/interaction/ (facets, a11y keyboard, canvas strata, hover/tooltip,
brush/zoom, point selection, lifecycle). Assertions preserved.
